After a ton of experimentation, I decided this is the best I can do with general scrollable elements. It is very difficult to get elements to be both scrollable and closeable dragged panes. It's potentially confusing though, too. So for now, scrollable content needs a close button (not yet added to the Mount Aubry balance pane).
Added a lot of the details content to Mount Aubry.

Mount Aubry could really benefit from a routing library to both get to deep content with a simple refresh and also to make open actions possible that require toggling multiple (potentially unknown) panes. For example, the "account summary" item in the side menu.